+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- add `fromCompressedPoints` method to `StealthAddress` trait
 - add `batches` method to `ClosableAsyncIterator`
 - remove `protocolWhitelistKey` fn after removing selectors from whitelist
 - add optional `throttleMs` field to `IterSyncOpts` and implement it in all sync adapters